Extract rise-backend-core: models, traits, quantity, state_machine
Scaffold the rise-backend-core crate as the dependency seam between
rise-deploy and the deployment backends (foundation for #377). This first
increment moves the shared contracts that have no database dependency:

- deployment-relevant data models (Deployment, Project, Environment,
  CustomDomain, env-var rows, and the status/termination enums)
- the DeploymentBackend trait + DeploymentUrls
- the RegistryProvider / EncryptionProvider provider traits and their
  credential types
- the pure quantity and state_machine helpers

rise-deploy re-exports every moved item from its original module path
(crate::db::models, controller, registry, encryption, deployment::quantity,
deployment::state_machine) so the rest of the backend compiles unchanged.
The crate is an optional, backend-feature-gated path dependency; the
default CLI build never references it.
