+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     ".",
     "crates/rise-backend-auth",
+    "crates/rise-backend-core",
     "crates/rise-resource-api",
     "crates/rise-resource-store",
     "crates/rise-runtime-sync",
@@ -69,6 +70,7 @@ cli = [
 backend = [
     # Server core (former "server" feature)
     "dep:rise-backend-auth", # strip-on-publish
+    "dep:rise-backend-core", # strip-on-publish
     "dep:rise-resource-api", # strip-on-publish
     "dep:rise-resource-store", # strip-on-publish
     "dep:rise-runtime-sync", # strip-on-publish
@@ -116,6 +118,7 @@ backend = [
 # --- BEGIN BACKEND (stripped before `cargo publish` in CI; see .github/workflows/ci.yml) ---
 # Workspace crates (backend only)
 rise-backend-auth = { path = "crates/rise-backend-auth", optional = true }
+rise-backend-core = { path = "crates/rise-backend-core", optional = true }
 rise-resource-api = { path = "crates/rise-resource-api", optional = true }
 rise-resource-store = { path = "crates/rise-resource-store", optional = true }
 rise-runtime-sync = { path = "crates/rise-runtime-sync", optional = true }
